Fixed bugs

+++ DrugProNET/QueryResult.aspx.cs
@@ -11,7 +11,7 @@ namespace DrugProNET
 {
     public partial class QueryResult : AdvertiseablePage
     {
-        private int interaction_distance;
+        private double interaction_distance;
         private bool protein_chain;
         private bool protein_atoms;
         private bool protein_residues;
@@ -30,7 +30,7 @@ namespace DrugProNET
 
             if (!IsPostBack)
             {
-                interaction_distance = int.Parse(Request.QueryString["interaction_distance"]);
+                interaction_distance = double.Parse(Request.QueryString["interaction_distance"]);
                 protein_chain = bool.Parse(Request.QueryString["protein_chain"]);
                 protein_atoms = bool.Parse(Request.QueryString["protein_atoms"]);
                 protein_residues = bool.Parse(Request.QueryString["protein_residues"]);
@@ -107,7 +107,7 @@ namespace DrugProNET
             }
         }
 
-        private void CreateInteractionList(PDB_Information PDB, int interaction_distance, bool protein_chain, bool protein_atoms, bool protein_residues, bool protein_residue_numbers, bool drug_atoms)
+        private void CreateInteractionList(PDB_Information PDB, double interaction_distance, bool protein_chain, bool protein_atoms, bool protein_residues, bool protein_residue_numbers, bool drug_atoms)
         {
             TableHeaderRow tableHeaderRow = new TableHeaderRow { TableSection = TableRowSection.TableHeader };
 
@@ -290,7 +290,7 @@ namespace DrugProNET
         protected void Download_List_Click(object sender, EventArgs e)
         {
             distances = (List<PDB_Distances>)Session["distances"];
-            interaction_distance = (int)Session["interaction_distance"];
+            interaction_distance = (double)Session["interaction_distance"];
             protein_chain = (bool)Session["protein_chain"];
             protein_atoms = (bool)Session["protein_atoms"];
             protein_residues = (bool)Session["protein_residues"];
